Work around weird issue where GDB was refusing to look up using the symbol table alone
This was causing our panic GDB command to decrease the stack pointer without jumping to the panic
function, thereby erroneously reporting 0 cleanups for all datapoints in builds without debug
symbols or those that had been optimized.  The issue also appeared to affect *certain* datapoints
in unoptimized builds having debug info: something about the depth of the call stack or its
particular contents would break the context for running the debug info search.

In both cases, the error was identifiable by removing the 2>&1 from the table script's loop guard
and looking for this error message in the output:
	No symbol 'panic' in current context

This patch fixes it by looking up the panic() function's address in the *symbol table* once at the
start, rather than relying on the high-level variable approach.