table: Fix subtle bug where the $frame GDB was not getting defined
We were getting lucky that the ni script was quitting GDB at the end of the realm() function, but
this wouldn't have worked with other functions.  The problem was revealed by running GDB standalone
and noticing that only the integer variables were printed at the start; asking about $frame revealed
that its value was void!