libunwind_a04279dd5736.patch: Patch for Rust 1.56.0's LLVM 13's libunwind to fix SIGABRTs
Injecting an exception just before an instruction that moves the stack pointer causes vanilla
libgcc_s or libunwind to skip applying the corresponding DW_CFA_def_cfa_offset "instruction" from
the DWARF .eh_frame section.  This patch alters the boundary case to fix the problem and eliminate
the SIGABRTs.  It requires building with -lunwind to override the _Unwind_* symbols from the (later)
implicit -lgcc_s, as well as overriding the search path with DT_RUNPATH or $LD_LIBRARY_PATH.