libunwind_a04279dd5736.patch: Update patch to handle the initial instruction of a function
The vanilla unwinder crashes on such instructions because it includes a hack that subtracts one from
the instruction pointer to guard against call instructions that may occur at the end of a function.
Unfortunately, this causes the pointer to be outside the function's address range, causing the
search for the exception-handling frame information to fail.  We work around this by skipping the
subtraction if the instruction pointer describes the very first byte of the function.