epilogue.s: Method and apparatus for fixing up execution state for cancellation during epilogues
Function epilogues are perilous because making calls during them can clobber stack-allocated
locals.  This code, if called just before the epilogue, performs the record keeping to allow us to
timetravel backward and simulate raising exceptions just before we entered the epilogue, even if it
had already started executing.

In order to work, the ingerc_epilogue_start() function will need CFI instructions so unwinding works
there.  The injected call to it needs to be followed by at least one instruction (even a nop is
fine) that precedes the epilogue so the timetravelled unwinding finds a landing pad.  And the
runtime injecting the exceptions needs to treat the jmp instruction at the end of
ingerc_epilogue_end() as if ingerc_epilogue_ra contained %rdi.